journal: specialize Timestamp
Change Timestamp from trait to a newtype wrapper of u64.

It turns out that a trait bound marker of Timestamp goes into premature optimization. Even a most complex clock consists of no more than three parts (seconds, nanoseconds, logical clock, all of which are integers).

We can later extend the Timestamp structure to simulate multiple parts, but specialize & simplify it for now.

This is a follow-up to #103.
This refers to #100.

Signed-off-by: tison <wander4096@gmail.com>