ci: fix boolean logic
This script is needed because GitHub Actions does't support path filter for jobs and steps.

The boolean logic should be, skip CI if and only if all paths-ignore matches and none paths matches.

Previously, the default value of paths is "*", which means all paths are matched. It makes paths-ignore ("*.md", "*.svg") takes no effect.

Sorry for introducing this boolean logic bug :-(

Signed-off-by: tison <wander4096@gmail.com>