server: fix segment fault (#660)
The address of the `addr` and `addrlen` can be changed when the listener is moved, which will cause segment fault.
Since we don't use those fields so far, we can simply remove them. If we need to get the socket address in the future, we can boxed the fields instead.